fix isuses with patcher
