+++ pyproject.toml
@@ -10,7 +10,6 @@ packages = [
 
 [tool.poetry.dependencies]
 python = ">=3.11,<4.0"
-playwright = "^1.48.0"
 clean-html-for-llm = "^1.3.2"
 webdriver-manager = "^4.0.2"
 python-dotenv = "^1.0.1"
